Fixed crash
Happend when a bow was deleted while a training was created with this bow
