+++ app/src/main/java/de/dreier/mytargets/features/training/edit/EditTrainingFragment.java
@@ -140,7 +140,7 @@ public class EditTrainingFragment extends EditFragmentBase implements DatePicker
         binding.bow.setOnUpdateListener(this::setScoringStyleForCompoundBow);
         binding.environment.setOnActivityResultContext(this);
         binding.trainingDate.setOnClickListener(view -> onDateClick());
-        
+
         if (trainingId == null) {
             ToolbarUtils.setTitle(this, R.string.new_training);
             binding.training.setText(getString(
